pyproject.toml
[project]
name = "llm-evaluation-ragas"
version = "0.1.0"
description = "RAGAS evaluation framework learning examples"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.4.0",
    "datasets>=4.4.2",
    "langchain-core>=1.2.6",
    "langchain-openai>=1.1.7",
    "litellm>=1.80.12",
    "openai>=2.14.0",
    "pymupdf>=1.26.7",
    "ragas>=0.4.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
]
